fix: save payment method whenever payment is completed
